Move hiffy helper functions out of cmd_hiffy
