+++ cmd/monorail/Cargo.toml
@@ -5,8 +5,6 @@ edition = "2021"
 description = "Management network control and debugging"
 
 [dependencies]
-cmd-hiffy.workspace = true
-
 hif.workspace = true
 idol.workspace = true
 vsc7448-info.workspace = true
