+++ cmd/net/Cargo.toml
@@ -7,8 +7,6 @@ description = "Management network device-side control and debugging"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-cmd-hiffy.workspace = true
-
 hif.workspace = true
 
 anyhow.workspace = true
