+++ cmd/powershelf/Cargo.toml
@@ -9,8 +9,6 @@ humility.workspace = true
 humility-cmd.workspace = true
 humility-hiffy.workspace = true
 humility-idol.workspace = true
-cmd-hiffy.workspace = true
-cmd-rpc.workspace = true
 pmbus.workspace = true
 
 hif.workspace = true
