+++ cmd/rendmp/Cargo.toml
@@ -16,7 +16,6 @@ num-traits.workspace = true
 num-derive.workspace = true
 zerocopy.workspace = true
 
-cmd-hiffy.workspace = true
 humility-cmd.workspace = true
 humility-hiffy.workspace = true
 humility-idol.workspace = true
