+++ cmd/update/Cargo.toml
@@ -11,7 +11,6 @@ clap.workspace = true
 parse_int.workspace = true
 log.workspace = true
 indicatif.workspace = true
-cmd-hiffy.workspace = true
 
 humility.workspace = true
 humility-cmd.workspace = true
